v1/requirements.txt
Django==1.11.3
gunicorn
whitenoise
psycopg2
djangorestframework
coreapi
Markdown
django-guardian
django-cors-headers
Pillow
cloudinary
dj-database-url==0.4.2
django-filter
